+++ src/net/sourceforge/fidocadj/graphic/nil/GraphicsNull.java
@@ -165,7 +165,7 @@ public class GraphicsNull implements GraphicsInterface
         @param isItalic true if an italic variant should be used.
         @param isBold true if a bold variant should be used.
     */
-    public void setFont(String name, int size, boolean isItalic,
+    public void setFont(String name, double size, boolean isItalic,
         boolean isBold)
     {
         /*Font f = new Font(name,
@@ -184,7 +184,7 @@ public class GraphicsNull implements GraphicsInterface
         @param name the name of the typeface.
         @param size the vertical size in pixels.
     */
-    public void setFont(String name, int size)
+    public void setFont(String name, double size)
     {
         setFont(name, size, false, false);
     }
